Move offset to pongon-only
When I did the cropping change I realized that "x" and "y" are obviously being passed to Pillow for the pasting, and it will explode if you pass it negatives. Since the offset is just for atlas convenience and pongon is the part that is atlas-related the offset is now there instead.